WA-390 Send rejectTransaction notification
