+++ app/routes/ConfirmTransaction/containers/ConfirmTransaction.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux'
 import EthUtil from 'ethereumjs-util'
 import BigNumber from 'bignumber.js'
 
-import { sendNotification } from 'utils/sendNotifications'
+import { sendNotification, requestConfirmationResponse } from 'utils/sendNotifications'
 import {
   getDecryptedEthAccount,
   createAccountFromMnemonic,
@@ -32,20 +32,17 @@ class ConfirmTransaction extends Component {
 
   showTransaction = (windowId) => {
     const { transactions } = this.props
-
     if (!transactions || transactions.length === 0) {
       return
     }
 
     const windowTransaction = transactions.filter(t => t.popupId === windowId)
     const hasTransaction = windowTransaction.length === 1 && windowTransaction[0].tx
-
     if (!hasTransaction) {
       return
     }
 
     const tx = windowTransaction[0].tx
-
     this.setState({
       transaction: {
         hash: tx.hash,
@@ -61,32 +58,32 @@ class ConfirmTransaction extends Component {
   handleConfirmTransaction = () => {
     const { transaction } = this.state
     const { seed, unlockedMnemonic } = this.props.account.secondFA
-
     const account = !unlockedMnemonic && this.password
       ? getDecryptedEthAccount(seed, this.password)
       : createAccountFromMnemonic(unlockedMnemonic)
-
-    console.log(account.getChecksumAddressString())
     const privateKey = account.getPrivateKey()
 
-    const owners = []
+    requestConfirmationResponse(
+      'confirmTransaction',
+      privateKey,
+      transaction.hash,
+    )
+  }
 
-    const signedTxHash = EthUtil.sha3(transaction.hash)
-    const vrsTxHash = EthUtil.ecsign(signedTxHash, privateKey)
-    const r = new BigNumber(EthUtil.bufferToHex(vrsTxHash.r))
-    const s = new BigNumber(EthUtil.bufferToHex(vrsTxHash.s))
-    const data = JSON.stringify({
-      type: 'confirmTransaction',
-      hash: transaction.hash,
-      r: r.toString(10),
-      s: s.toString(10),
-      v: vrsTxHash.v.toString(10),
-    })
+  handleRejectTransaction = () => {
+    const { transaction } = this.state
+    const { seed, unlockedMnemonic } = this.props.account.secondFA
+    const account = !unlockedMnemonic && this.password
+      ? getDecryptedEthAccount(seed, this.password)
+      : createAccountFromMnemonic(unlockedMnemonic)
+    const privateKey = account.getPrivateKey()
 
-    const response = sendNotification(owners, data, privateKey)
-    if (response) {
-      console.log(response.status)
-    }
+    requestConfirmationResponse(
+      'rejectTransaction',
+      privateKey,
+      transaction.hash,
+      'GNO'
+    )
   }
 
   render() {
@@ -96,6 +93,7 @@ class ConfirmTransaction extends Component {
       <Layout
         transaction={transaction}
         handleConfirmTransaction={this.handleConfirmTransaction}
+        handleRejectTransaction={this.handleRejectTransaction}
       />
     )
   }
